Move more service code into service/http/
 * the service HTTP adapter (http/server -> service/http)
 * notifications code (notifications -> service/notifications)
 * the service server (server -> service/server)
